backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6

sqlalchemy==2.0.30
pydantic==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.18
aiofiles==23.2.1

httpx==0.27.0
requests==2.32.3

geopy==2.4.1
shapely==2.0.4
folium==0.17.0

numpy==2.2.0
pandas==2.2.2
Pillow==10.4.0
scikit-learn==1.5.1
matplotlib==3.9.0

torch==2.2.2
torchvision==0.17.2
ultralytics==8.2.0



reportlab==4.2.2
roboflow==1.1.30